+++ src/datasources/asset-calibration/components/AssetCalibrationQueryBuilder.test.tsx
@@ -2,6 +2,7 @@ import React, { ReactNode } from "react";
 import { AssetCalibrationQueryBuilder } from "./AssetCalibrationQueryBuilder";
 import { render } from "@testing-library/react";
 import { Workspace } from "core/types";
+import { SystemMetadata } from "datasources/system/types";
 
 describe('AssetCalibrationQueryBuilder', () => {
   describe('useEffects', () => {
@@ -9,8 +10,8 @@ describe('AssetCalibrationQueryBuilder', () => {
 
     const containerClass = 'smart-filter-group-condition-container'
 
-    function renderElement(workspaces: Workspace[], filter?: string) {
-      reactNode = React.createElement(AssetCalibrationQueryBuilder, { workspaces, filter, onChange: jest.fn() });
+    function renderElement(workspaces: Workspace[], systems: SystemMetadata[], filter?: string) {
+      reactNode = React.createElement(AssetCalibrationQueryBuilder, { workspaces, systems, filter, onChange: jest.fn(), areDependenciesLoaded: true });
       const renderResult = render(reactNode);
       return {
         renderResult,
@@ -19,18 +20,29 @@ describe('AssetCalibrationQueryBuilder', () => {
     }
 
     it('should render empty query builder', () => {
-      const { renderResult, conditionsContainer } = renderElement([], '');
+      const { renderResult, conditionsContainer } = renderElement([], [], '');
       expect(conditionsContainer.length).toBe(1);
       expect(renderResult.findByLabelText('Empty condition row')).toBeTruthy();
     })
 
-    it('should populate query builder', () => {
-      const workspace = { id: '1', name: 'Selected workspace' } as Workspace
-      const { conditionsContainer } = renderElement([workspace], 'Workspace = "1" && ModelName = "SomeRandomModelName"');
+    it('should select workspace in query builder', () => {
+      const workspace = { id: '1', name: 'Selected workspace' } as Workspace;
+      const system = { id: '1', alias: 'Selected system' } as SystemMetadata;
+      const { conditionsContainer } = renderElement([workspace], [system], 'Workspace = "1" && ModelName = "SomeRandomModelName"');
 
       expect(conditionsContainer?.length).toBe(2);
       expect(conditionsContainer.item(0)?.textContent).toContain(workspace.name);
       expect(conditionsContainer.item(1)?.textContent).toContain("SomeRandomModelName");
     })
-  })
-})
+
+    it('should select system in query builder', () => {
+      const workspace = { id: '1', name: 'Selected workspace' } as Workspace;
+      const system = { id: '1', alias: 'Selected system' } as SystemMetadata;
+
+      const { conditionsContainer } = renderElement([workspace], [system], 'Location = "1"');
+
+      expect(conditionsContainer?.length).toBe(1);
+      expect(conditionsContainer.item(0)?.textContent).toContain(system.alias);
+    });
+  });
+});