+++ src/datasources/asset-calibration/constants.ts
@@ -1,19 +1,30 @@
 import { QueryBuilderOperations } from "core/query-builder.constants";
 import { AssetTypeOptions, BusTypeOptions, QBField } from "./types";
 
+export enum AssetCalibrationFieldNames {
+    LOCATION = 'Location',
+    WORKSPACE = 'Workspace',
+    MODEL_NAME = 'ModelName',
+    VENDOR_NAME = 'VendorName',
+    BUS_TYPE = 'BusType',
+    ASSET_TYPE = 'AssetType',
+}
+
 export const AssetCalibrationFields: Record<string, QBField> = {
     LOCATION: {
         label: 'Location',
-        dataField: 'Location',
+        dataField: AssetCalibrationFieldNames.LOCATION,
         filterOperations: [
             QueryBuilderOperations.EQUALS.name,
             QueryBuilderOperations.DOES_NOT_EQUAL.name
         ],
+        lookup: {
+            dataSource: []
+        }
     },
     WORKSPACE: {
         label: 'Workspace',
-        dataField: 'Workspace',
-        dataType: 'string',
+        dataField: AssetCalibrationFieldNames.WORKSPACE,
         filterOperations: [
             QueryBuilderOperations.EQUALS.name,
             QueryBuilderOperations.DOES_NOT_EQUAL.name
@@ -24,61 +35,50 @@ export const AssetCalibrationFields: Record<string, QBField> = {
     },
     MODEL_NAME: {
         label: 'Model Name',
-        dataField: 'ModelName',
+        dataField: AssetCalibrationFieldNames.MODEL_NAME,
         filterOperations: [
             QueryBuilderOperations.EQUALS.name,
-            QueryBuilderOperations.DOES_NOT_EQUAL.name,
-            QueryBuilderOperations.CONTAINS.name,
-            QueryBuilderOperations.DOES_NOT_CONTAIN.name
+            QueryBuilderOperations.DOES_NOT_EQUAL.name
         ],
     },
     VENDOR_NAME: {
         label: 'Vendor Name',
-        dataField: 'VendorName',
+        dataField: AssetCalibrationFieldNames.VENDOR_NAME,
         filterOperations: [
             QueryBuilderOperations.EQUALS.name,
-            QueryBuilderOperations.DOES_NOT_EQUAL.name,
-            QueryBuilderOperations.CONTAINS.name,
-            QueryBuilderOperations.DOES_NOT_CONTAIN.name
+            QueryBuilderOperations.DOES_NOT_EQUAL.name
         ],
     },
     BUS_TYPE: {
         label: 'Bus Type',
-        dataField: 'BusType',
+        dataField: AssetCalibrationFieldNames.BUS_TYPE,
         filterOperations: [
             QueryBuilderOperations.EQUALS.name,
             QueryBuilderOperations.DOES_NOT_EQUAL.name
         ],
         lookup: {
             dataSource: BusTypeOptions,
-            readonly: true,
         }
     },
     ASSET_TYPE: {
         label: 'Asset Type',
-        dataField: 'AssetType',
+        dataField: AssetCalibrationFieldNames.ASSET_TYPE,
         filterOperations: [
             QueryBuilderOperations.EQUALS.name,
             QueryBuilderOperations.DOES_NOT_EQUAL.name
         ],
         lookup: {
             dataSource: AssetTypeOptions,
-            readonly: true,
         },
     },
 };
 
 export const AssetCalibrationStaticFields = [
-    AssetCalibrationFields.LOCATION,
     AssetCalibrationFields.MODEL_NAME,
     AssetCalibrationFields.VENDOR_NAME,
     AssetCalibrationFields.ASSET_TYPE,
     AssetCalibrationFields.BUS_TYPE
 ];
-
-export const AssetComputedDataFields = {
-    'Location': '(Location.PhysicalLocation = "{value}" || Location.MinionId = "{value}")',
-}
 
 export const assetSummaryFields = {
     TOTAL: 'Total',