+++ src/datasources/asset/module.ts
@@ -1,8 +1,10 @@
 import { DataSourcePlugin } from '@grafana/data';
 import { AssetDataSource } from './AssetDataSource';
 import { AssetQueryEditor } from './components/AssetQueryEditor';
-import { HttpConfigEditor } from 'core/components/HttpConfigEditor';
+import { AssetDataSourceOptions, AssetQuery } from './types/types';
+import { AssetConfigEditor } from './AssetConfigEditor';
 
-export const plugin = new DataSourcePlugin(AssetDataSource)
-  .setConfigEditor(HttpConfigEditor)
+
+export const plugin = new DataSourcePlugin<AssetDataSource, AssetQuery, AssetDataSourceOptions>(AssetDataSource)
+  .setConfigEditor(AssetConfigEditor)
   .setQueryEditor(AssetQueryEditor);